chore: moves resources, module mocking and docs to tests
