+++ e2e/specs/confirmations/regression/new-networks-send-erc721.spec.ts
@@ -8,7 +8,7 @@ import { SMART_CONTRACTS } from '../../../../app/util/test/smart-contracts';
 import { ActivitiesViewSelectorsText } from '../../../../app/components/Views/ActivityView/ActivitiesView.testIds';
 import Assertions from '../../../../tests/framework/Assertions';
 import { buildPermissions } from '../../../../tests/framework/fixtures/FixtureUtils';
-import { NETWORK_TEST_CONFIGS } from '../../../resources/mock-configs';
+import { NETWORK_TEST_CONFIGS } from '../../../../tests/resources/mock-configs';
 import { DappVariants } from '../../../../tests/framework/Constants';
 import TestHelpers from '../../../helpers';
