+++ e2e/specs/quarantine/deeplink-to-buy-flow.failing.ts
@@ -8,7 +8,7 @@ import BuyGetStartedView from '../../pages/Ramps/BuyGetStartedView';
 import BuildQuoteView from '../../pages/Ramps/BuildQuoteView';
 import TokenSelectBottomSheet from '../../pages/Ramps/TokenSelectBottomSheet';
 import Assertions from '../../../tests/framework/Assertions';
-import { PopularNetworksList } from '../../resources/networks.e2e';
+import { PopularNetworksList } from '../../../tests/resources/networks.e2e';
 import NetworkEducationModal from '../../pages/Network/NetworkEducationModal';
 
 // This test was migrated to the new framework but should be reworked to use withFixtures properly
