+++ e2e/specs/quarantine/deeplinks.failing.ts
@@ -15,7 +15,7 @@ import { importWalletWithRecoveryPhrase } from '../../viewHelper';
 import Accounts from '../../../wdio/helpers/Accounts';
 import TabBarComponent from '../../pages/wallet/TabBarComponent';
 import Assertions from '../../../tests/framework/Assertions';
-import { PopularNetworksList } from '../../resources/networks.e2e';
+import { PopularNetworksList } from '../../../tests/resources/networks.e2e';
 
 //const BINANCE_RPC_URL = 'https://bsc-dataseed1.binance.org';
