+++ e2e/specs/quarantine/onramp.failing.ts
@@ -3,7 +3,7 @@ import WalletView from '../../pages/wallet/WalletView';
 import FundActionMenu from '../../pages/UI/FundActionMenu';
 import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
 import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import { CustomNetworks } from '../../resources/networks.e2e';
+import { CustomNetworks } from '../../../tests/resources/networks.e2e';
 import { SmokeTrade } from '../../tags';
 import Assertions from '../../../tests/framework/Assertions';
 import BuildQuoteView from '../../pages/Ramps/BuildQuoteView';
