+++ e2e/specs/quarantine/permission-system-remove.failing.ts
@@ -8,7 +8,7 @@ import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
 import { loginToApp, navigateToBrowserView } from '../../viewHelper';
 import Assertions from '../../../tests/framework/Assertions';
 
-import { PopularNetworksList } from '../../resources/networks.e2e';
+import { PopularNetworksList } from '../../../tests/resources/networks.e2e';
 
 import WalletView from '../../pages/wallet/WalletView';
 import NetworkListModal from '../../pages/Network/NetworkListModal';
