+++ e2e/specs/wallet/balance-empty-state.spec.ts
@@ -7,7 +7,7 @@ import NetworkListModal from '../../pages/Network/NetworkListModal';
 import NetworkEducationModal from '../../pages/Network/NetworkEducationModal';
 import NetworkManager from '../../pages/wallet/NetworkManager';
 import Assertions from '../../../tests/framework/Assertions';
-import { CustomNetworks } from '../../resources/networks.e2e';
+import { CustomNetworks } from '../../../tests/resources/networks.e2e';
 import BuyGetStartedView from '../../pages/Ramps/BuyGetStartedView';
 import { NetworkToCaipChainId } from '../../../app/components/UI/NetworkMultiSelector/NetworkMultiSelector.constants';
