+++ tests/module-mocking/sentry/core.ts
