+++ tests/module-mocking/sentry/react-native.ts
