+++ tests/resources/collectibles.json
